CLI: Print unknown argument errors
All other validation errors are handled and printed in cliApp.ts:83.
These errors were previously not printed and silently returned to the
user on the error channel. We should either print all validation errors
or handle all of them silently and leave printing/reporting to the
caller. This commit implements the former option.