fix(ConfigProvider.fromEnv): use import.meta.env instead of process.env
The reason for this is that this allows using environment variables in the browser with vite. Vite has a feature where all environment variables that have a VITE_ prefix are exposed in import.meta.env. And even in node.js import.meta.env is a alias to process.env so i don't see any reason why ConfigProvider.fromEnv only supported process.env