Fix the incorrect TypeScript type of httpApp (partially)
The HttpRouter.toHttpApp returns an App.Default<E | RouteNotFound>,
but that type gets suppressed by the input type of Router.router, which
has `unknown` for the `E`, so the whole thing become `unknown`.

Additionally, the `apiMiddleware` has the type:

    <E, R>(self: App.Default<E, R>): App.Default<any, any>

So once again, the RouteNotFound is lost when the router is piped
through the middleware.

Both of these places where type information is lost could have led to
the following oversight:

The error is eventually captured by catchAllCause, and an attempt is
made to encode it. We don't know if the encoding succeeds, but by
default it will not succeed for RouteNotFound, causing it to be
re-failed to the end user. It is then likely caught by the
`withErrorHandling` part of `HttpApp.toHandled` which uses
`ServerError.causeResponse` to convert it to a Response.

However, in userland it's not clear why this is happening, because the
return type from `httpApp` states that the created HttpApp cannot fail.

This commit adds RouteNotFound to the error branch of the returned
HttpApp, making it clearer for users how RouteNotFound is propegated and
handled, and allowing users to add their own handlers.

------------------------------------------------------------------------

Note:

Technically, all errors that could occur according to the injected HttpApi
should be included in the output type of the httpApp. We never know for
sure that the encoder will succeed, and so we could always re-fail them.

The thing is, we can't know these in the `httpApp` constructor, because
the `HttpApi` is injected via the `HttpApi.Api` tag, which loses the `E`
type on the `HttpApi` it contains. So I left this out of scope for this
fix, because it would require a lot of refactoring.